+++ constants.js
@@ -1,0 +1,9 @@
+const DEFAULT_RESTAURANT_CHOICES = {
+  cuisines: [],
+  dishCategories: [],
+  dishNames: [],
+};
+
+module.exports = {
+  DEFAULT_RESTAURANT_CHOICES,
+};
