+++ older_code/dumm_data.js
@@ -1,71 +1,69 @@
 const { RestaurantModel } = require("./models")
 
 const dummyDishes = [
-    {
-        name: "Mozzarella E Caprese",
-        description: "Plum tomato and buffalo mozzarella, basil pesto, pine nuts, extra virgin olive oil",
-        price: "395",
-        category: "INSALATA / SALADS",
-        dishType: "veg",
+  {
+    name: "Mozzarella E Caprese",
+    description:
+      "Plum tomato and buffalo mozzarella, basil pesto, pine nuts, extra virgin olive oil",
+    price: "395",
+    category: "INSALATA / SALADS",
+    dishType: "Veg",
+  },
+  {
+    name: "Insalata Di Caesar",
+    description:
+      "Ice burg and romaine lettuce, herbed chicken, garlic croutons, and parmesan cheese, poached egg",
+    price: "395",
+    category: "INSALATA / SALADS",
+    dishType: "Veg",
+  },
+  {
+    name: "Spaghetti Aglio Olid Gamberi",
+    description: "Sphagetti tossed in olive oil, garlic, chilli, herbs etc.",
+    price: "535",
+    category: "PASTA E RISOTTI",
+    dishType: "Non_veg",
+  },
+  {
+    name: "Penne Milanese",
+    description: "Penne with chicken, mushroom, oregano etc.",
+    price: "485",
+    category: "PASTA E RISOTTI",
+    dishType: "Non_veg",
+    specialTag: "Bestseller",
+  },
 
-    },
-    {
-        name: "Insalata Di Caesar",
-        description: "Ice burg and romaine lettuce, herbed chicken, garlic croutons, and parmesan cheese, poached egg",
-        price: "395",
-        category: "INSALATA / SALADS",
-        dishType: "veg",
-
-    },
-    {
-        name: "Spaghetti Aglio Olid Gamberi",
-        description: "Sphagetti tossed in olive oil, garlic, chilli, herbs etc.",
-        price: "535",
-        category: "PASTA E RISOTTI",
-        dishType: "non_veg",
-
-    },
-    {
-        name: "Penne Milanese",
-        description: "Penne with chicken, mushroom, oregano etc.",
-        price: "485",
-        category: "PASTA E RISOTTI",
-        dishType: "non_veg",
-        specialTag: "Bestseller",
-    },
-
-    {
-        name: "Salmone",
-        description: "Tomato sauce, mozzarella, smoked salmon, onion, capers",
-        price: "655",
-        category: "SELECTION OF PIZZA",
-        dishType: "non_veg",
-
-    },
-    {
-        name: "Pollo Funghi Spinaci",
-        description: "Tomato sauce, mozzarella, pesto chicken, spinach, olives, sun-dried tomatoes",
-        price: "585",
-        category: "SELECTION OF PIZZA",
-        dishType: "non_veg",
-
-    },
-    {
-        name: "Tiramisu",
-        description: "Italian mascarpone, lavazza espresso, coffee ice cream",
-        price: "295",
-        category: "DOLCI / DESSERTS",
-        dishType: "veg",
-        specialTag: "Bestseller",
-    },
-    {
-        name: "Orange Creme Bruilee",
-        description: "Classcic creme brulee with orange zest",
-        price: "295",
-        category: "DOLCI / DESSERTS",
-        dishType: "veg",
-    },
-]
+  {
+    name: "Salmone",
+    description: "Tomato sauce, mozzarella, smoked salmon, onion, capers",
+    price: "655",
+    category: "SELECTION OF PIZZA",
+    dishType: "Non_veg",
+  },
+  {
+    name: "Pollo Funghi Spinaci",
+    description:
+      "Tomato sauce, mozzarella, pesto chicken, spinach, olives, sun-dried tomatoes",
+    price: "585",
+    category: "SELECTION OF PIZZA",
+    dishType: "Non_veg",
+  },
+  {
+    name: "Tiramisu",
+    description: "Italian mascarpone, lavazza espresso, coffee ice cream",
+    price: "295",
+    category: "DOLCI / DESSERTS",
+    dishType: "Veg",
+    specialTag: "Bestseller",
+  },
+  {
+    name: "Orange Creme Bruilee",
+    description: "Classcic creme brulee with orange zest",
+    price: "295",
+    category: "DOLCI / DESSERTS",
+    dishType: "Veg",
+  },
+];
 
 const location ={ 
     latitude: 12.9143474, longitude: 77.6747474, areaName: "Sarjapur",
